requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
astroid==2.7.2
beautifulsoup4==4.9.3
cffi==1.14.6
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pkg_resources==0.0.0
platformdirs==2.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pylint==2.10.2
python-dotenv==0.19.0
pytz==2021.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
